# Conflicts:
#	src/PrivateKey.cpp
#	src/PrivateKey.h

+++ wasm/src/keystore/default-impl.ts
@@ -82,7 +82,8 @@ export class Default implements Types.IKeyStore {
     name: string,
     password: string,
     coin: CoinType,
-    encryption: StoredKeyEncryption
+    encryption: StoredKeyEncryption,
+    derivation: Derivation
   ): Promise<Types.Wallet> {
     return new Promise((resolve, reject) => {
       const { StoredKey, PrivateKey, Curve, StoredKeyEncryption } = this.core;
@@ -95,7 +96,28 @@ export class Default implements Types.IKeyStore {
         throw Types.Error.InvalidKey;
       }
       let pass = Buffer.from(password);
-      let storedKey = StoredKey.importPrivateKeyWithEncryption(key, name, pass, coin, encryption);
+      let storedKey = StoredKey.importPrivateKeyWithEncryptionAndDerivation(key, name, pass, coin, encryption, derivation);
+      let wallet = this.mapWallet(storedKey);
+      storedKey.delete();
+      this.importWallet(wallet)
+        .then(() => resolve(wallet))
+        .catch((error) => reject(error));
+    });
+  }
+
+  importKeyEncoded(
+    key: string,
+    name: string,
+    password: string,
+    coin: CoinType,
+    encryption: StoredKeyEncryption,
+    derivation: Derivation
+  ): Promise<Types.Wallet> {
+    return new Promise((resolve, reject) => {
+      const { StoredKey, PrivateKey, Curve, StoredKeyEncryption } = this.core;
+
+      let pass = Buffer.from(password);
+      let storedKey = StoredKey.importPrivateKeyEncodedWithEncryptionAndDerivation(key, name, pass, coin, encryption, derivation);
       let wallet = this.mapWallet(storedKey);
       storedKey.delete();
       this.importWallet(wallet)
@@ -160,6 +182,46 @@ export class Default implements Types.IKeyStore {
         default:
           throw Types.Error.InvalidJSON;
       }
+      storedKey.delete();
+      return value;
+    });
+  }
+
+  getWalletType(id: string): Promise<Types.WalletType> {
+    return this.load(id).then((wallet) => wallet.type);
+  }
+
+  exportMnemonic(id: string, password: string): Promise<string> {
+    return this.load(id).then((wallet) => {
+      if (wallet.type !== Types.WalletType.Mnemonic) {
+        throw Types.Error.UnsupportedWalletType;
+      }
+      let storedKey = this.mapStoredKey(wallet);
+      let value = storedKey.decryptMnemonic(Buffer.from(password));
+      storedKey.delete();
+      return value;
+    });
+  }
+
+  exportPrivateKey(id: string, password: string): Promise<Uint8Array> {
+    return this.load(id).then((wallet) => {
+      if (wallet.type !== Types.WalletType.PrivateKey) {
+        throw Types.Error.UnsupportedWalletType;
+      }
+      let storedKey = this.mapStoredKey(wallet);
+      let value = storedKey.decryptPrivateKey(Buffer.from(password));
+      storedKey.delete();
+      return value;
+    });
+  }
+
+  exportPrivateKeyEncoded(id: string, password: string): Promise<string> {
+    return this.load(id).then((wallet) => {
+      if (wallet.type !== Types.WalletType.PrivateKey) {
+        throw Types.Error.UnsupportedWalletType;
+      }
+      let storedKey = this.mapStoredKey(wallet);
+      let value = storedKey.decryptPrivateKeyEncoded(Buffer.from(password));
       storedKey.delete();
       return value;
     });
